Ailchemy Documents
These lexicons are my language that I’ve been putting together. Along with the starter guide, each one needs to be taken with a grain of salt.

This is my way; this isn't "the" way. They will just help you find your way in the dark, in a "Survival Guides" style.